create file structure and add required dependensies
