+++ .gitignore
@@ -1,24 +1,41 @@
-# Logs
-logs
-*.log
+# dependencies
+/node_modules
+/project.inlang
+/extractedTranslations
+/storybook-static
+/.pnp
+.pnp.js
+.yarn/install-state.gz
+
+# testing
+/coverage
+
+# production
+/build
+
+# misc
+.DS_Store
+*.pem
+.vscode
+
+# debug
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
-pnpm-debug.log*
-lerna-debug.log*
 
-node_modules
-dist
-dist-ssr
-*.local
+# local env files
+.env*.local
+.env
 
-# Editor directories and files
-.vscode/*
-!.vscode/extensions.json
-.idea
-.DS_Store
-*.suo
-*.ntvs*
-*.njsproj
-*.sln
-*.sw?
+# vercel
+.vercel
+
+# Log Files
+*.log
+
+# XML Files
+*.xml
+
+# typescript
+*.tsbuildinfo
+*storybook.log
