+++ public/index.html
@@ -3,9 +3,9 @@
 
   <head>
     <meta charset="UTF-8" />
-    <link rel="icon" type="image/svg+xml" href="" />
+    <link rel="icon" type="image/svg+xml" href="/vite.svg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>PeakVibe</title>
+    <title>Vite + React + TS</title>
   </head>
 
   <body>
